Profile display added, Profile Avatar added everything working
