+++ src/pages/AllPosts.jsx
@@ -10,44 +10,7 @@ function AllPosts() {
         setPosts(posts.documents)
       }
     })
-    const userData = useSelector((state) => state.auth.userData);
-    const [profileChecked, setProfileChecked] = useState(false);
  
-   
-    
-    useEffect(() => {
-        const checkAndCreateProfile = async () => {
-            if (userData && !profileChecked) {
-                try {
-                    const profile = await appwriteService.getProfile(userData.$id);
-                    if (!profile) {
-                        // If profile doesn't exist, create one
-                        await appwriteService.createProfile({
-                            username: userData.name,
-                            bio: 'Welcome to my profile',
-                            profilepic: '',
-                            userID: userData.$id
-                        });
-                    }
-                    setProfileChecked(true);
-                } catch (error) {
-                    if (error.message === "Profile not found") {
-                        await appwriteService.createProfile({
-                            username: userData.name,
-                            bio: 'Welcome to my profile',
-                            profilepic: '',
-                            userID: userData.$id
-                        });
-                    } else {
-                        console.error('Failed to fetch or create profile', error);
-                    }
-                    setProfileChecked(true); // Ensure this runs only once
-                }
-            }
-        };
-
-        checkAndCreateProfile();
-    }, [userData, profileChecked]);
     
   return (
     <div className='w-full py-8'>
