+++ src/pages/UserProfile.jsx
@@ -10,6 +10,44 @@ export default function UserProfile() {
     const [profile, setProfile] = useState(null);
     const userData = useSelector((state) => state.auth.userData);
     const { userId } = useParams();
+    const [profileChecked, setProfileChecked] = useState(false);
+ 
+   
+    
+    useEffect(() => {
+        const checkAndCreateProfile = async () => {
+            if (userData && !profileChecked) {
+                try {
+                    const profile = await appwriteService.getProfile(userData.$id);
+                    if (!profile) {
+                        // If profile doesn't exist, create one
+                        await appwriteService.createProfile({
+                            username: userData.name,
+                            bio: 'Welcome to my profile',
+                            profilepic: '',
+                            userID: userData.$id
+                        });
+                    }
+                    setProfileChecked(true);
+                } catch (error) {
+                    if (error.message === "Profile not found") {
+                        await appwriteService.createProfile({
+                            username: userData.name,
+                            bio: 'Welcome to my profile',
+                            profilepic: '',
+                            userID: userData.$id
+                        });
+                    } else {
+                        console.error('Failed to fetch or create profile', error);
+                    }
+                    setProfileChecked(true); // Ensure this runs only once
+                }
+            }
+        };
+
+        checkAndCreateProfile();
+    }, [userData, profileChecked]);
+    
 
     useEffect(() => {
         const fetchUserProfile = async () => {
